amstex.sh was still using {ini,vir]tex superseded by {ini,vir}prote.
The recipe was succeeding on some installations and failing on others.
The reason was that initex(1) and virtex(1) were still used, that may
exist on older kerTeX installations; but on new ones, Prote, an
extension on TeX and e-Tex is used and initex(1) and virtex(1) don't
exist, replaced by iniprote(1) and virprote(1) resp.